isolate json generation to sepparate activity
